refactor(relay): split the artifact handler along its two audiences
max-lines caught server.ts and relay.ts, and the fix is the split the file
was already asking for rather than a bigger allowance.

The public page moves to published-page.ts. Everything the API does happens
for an authenticated account; everything the page does happens for a stranger
following a link, with no credential and no session, over bytes somebody else
wrote. Apart, it is hard to reach for a session in the one place there is
deliberately never going to be one.

Reading and validating a write moves to request-body.ts — the one part that
touches bytes a stranger chose, easier to argue about on its own.

The wiring moves to wiring.ts, where "no artifacts at all" is one branch in
one place, so the off path visibly allocates nothing and serves nothing.
